added some integer tests
