+++ Cargo.toml
@@ -129,7 +129,7 @@ enum_dispatch = "0.3.13"
 flate2 = "1.1.9"
 heck = "0.5"
 hmac = "=0.13.0"
-indexmap = "2.14"
+indexmap = "2"
 itertools = "0.14.0"
 libloading = "0.9"
 lru = "0.18.0"
