+++ pumpkin/Cargo.toml
@@ -108,6 +108,7 @@ postcard = { workspace = true }
 tracing-serde-structured = { workspace = true }
 
 bitflags = { workspace = true }
+indexmap.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
