+++ pumpkin/src/command/suggestion/mod.rs
@@ -381,8 +381,8 @@ mod test {
         assert_eq!(
             &merged.suggestions,
             &[
-                Suggestion::without_tooltip(StringRange::between(4, 5), "BAR"),
                 Suggestion::without_tooltip(StringRange::between(4, 5), "bar"),
+                Suggestion::without_tooltip(StringRange::between(4, 5), "BAR"),
                 Suggestion::without_tooltip(StringRange::between(4, 5), "bars"),
                 Suggestion::without_tooltip(StringRange::between(4, 5), "baz"),
                 Suggestion::without_tooltip(StringRange::between(4, 5), "foo"),
